Rename serial to serialnumber
